fix(api): prevent deletion of import files on error to allow retries

+++ api/src/prescription/learner-management/domain/usecases/import-learners/import-learners-from-fregata-file.js
@@ -57,7 +57,9 @@ const importLearnersFromFregataFile = withTransaction(async function ({
   } finally {
     organizationImport.process({ errors });
     await organizationImportRepository.save(organizationImport);
-    await importStorage.deleteFile({ filename: organizationImport.filename });
+    if (errors.length === 0) {
+      await importStorage.deleteFile({ filename: organizationImport.filename });
+    }
   }
 });
