+++ api/src/prescription/learner-management/domain/usecases/import-learners/import-learners-from-generic-file.js
@@ -66,7 +66,9 @@ const importLearnersFromGenericFile = async function ({
   } finally {
     organizationImport.process({ errors });
     await organizationImportRepository.save(organizationImport);
-    await importStorage.deleteFile({ filename: organizationImport.filename });
+    if (errors.length === 0) {
+      await importStorage.deleteFile({ filename: organizationImport.filename });
+    }
   }
 };
